Fixed regression in ArrayWidget. Wrote unit test to prevent the bug from croping up again.
-Kurtis

+++ src/Optika_SpinBoxApplier.hpp
@@ -211,7 +211,7 @@ public:
 			spinBox->setMinimum(EnhancedNumberTraits<double>::min());
 			spinBox->setMaximum(EnhancedNumberTraits<double>::max());
 			spinBox->setSingleStep(EnhancedNumberTraits<double>::defaultStep());
-			spinBox->setDecimals(EnhancedNumberTraits<float>::defaultPrecision());
+			spinBox->setDecimals(EnhancedNumberTraits<double>::defaultPrecision());
 		}
 	}
